Separate processors by concern, move scoping to new module
