+++ src/floorplan_dsl/processors/validation/variation.py
